+++ app/src/main/java/com/ejooyoung/pdf_reader/database/model/Contents.kt
@@ -3,5 +3,5 @@ package com.ejooyoung.pdf_reader.database.model
 interface Contents {
     fun getGuid(): String
     fun getTitle(): String
-    fun getPageIdx(): Long
+    fun getPageIdx(): Int
 }
